pyproject.toml
[tool.poetry]
name = "repl-nix-doomberg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_doomberg"}]

[tool.poetry.dependencies]
python = "3.12.3"
dash = "^2.17.1"
pandas = "^2.2.2"
langchain-openai = "^0.1.14"
python-dotenv = "^1.0.1"
crewai = "^0.36.0"
langchain-community = "^0.2.6"
langchain-mistralai = "^0.1.9"
langchain = "^0.2.7"
openai = "^1.35.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
